Applied patch for #3687.
git-svn-id: http://svn.osgeo.org/mapserver/trunk@11010 7532c77e-422f-0410-93f4-f0b67bdd69e2

+++ mapwcs.h
@@ -156,7 +156,7 @@ typedef wcs20SubsetObj * wcs20SubsetObjPtr;
 typedef struct
 {
     char *name;         /* name of the axis */
-    long size;          /* pixelsize of the axis */
+    int size;           /* pixelsize of the axis */
     double resolution;  /* resolution of the axis */
     char *resolutionUOM; /* resolution units of measure */
     wcs20SubsetObjPtr subset;
